src/main/java/com/daniboy/pageobjects/store/components/Product.java
package com.daniboy.pageobjects.store.components;

public class Product {
    private String name;
    private double price;
    private int quantity = 1;

    public Product(String name, double price) {
        this.name = name;
        this.price = price;
    }

    public String getName() {
        return name;
    }

    public double getPrice() {
        return price;
    }

    public int getQuantity() {
        return this.quantity;
    }
    public void setQuantity(int quantity) {
        this.quantity = quantity;
    }
}
